+++ website/src/examples/progress_indeterminate.rs
@@ -1,0 +1,7 @@
+use leptos::*;
+use leptos_bulma::elements::BProgress;
+
+#[component]
+pub fn ProgressIndeterminate() -> impl IntoView {
+    view! { <BProgress max=100 /> }
+}