+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Progress bar element (`BProgress`).
+
 ## [0.5.0] - 2024-08-23
 
 ### Changed